% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_residualized_pair.R
\name{compute_residualized_pair}
\alias{compute_residualized_pair}
\title{Compute residualized response and exposure}
\usage{
compute_residualized_pair(Y, A, EY_C, EA_C)
}
\arguments{
\item{Y}{Numeric outcome vector.}

\item{A}{Numeric matrix or data frame of observed exposures.}

\item{EY_C}{Numeric vector of predictions for \code{E[Y | C]}.}

\item{EA_C}{Numeric matrix of predictions for \code{E[A | C]}.}
}
\value{
A list with \code{Y_tilde} and \code{A_tilde}.
}
\description{
Compute residualized response and exposure
}
