% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{csdr_default_sl_library}
\alias{csdr_default_sl_library}
\title{Default SuperLearner regression library for CSDR}
\usage{
csdr_default_sl_library()
}
\value{
A character vector of SuperLearner wrapper names.
}
\description{
Default SuperLearner regression library for CSDR
}
