% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_pseudo_outcomes.R
\name{estimate_pseudo_outcomes}
\alias{estimate_pseudo_outcomes}
\title{Compute Doubly Robust Pseudo-Outcomes from Kennedy et al. (2017) JRSS-B}
\usage{
estimate_pseudo_outcomes(Y, X, C, out_model, gps_model, delta_n = 1e-16)
}
\arguments{
\item{Y}{Numeric vector of observed outcomes (length n).}

\item{X}{Numeric matrix or data frame of observed treatments (n x p).}

\item{C}{Numeric matrix or data frame of observed confounders (n x q).}

\item{out_model}{An S3 object of class "outcome_model".}

\item{gps_model}{An S3 object representing a global conditional density model.}

\item{delta_n}{Small positive threshold to floor propensity scores. Defaults to 1e-16.}
}
\value{
A numeric vector of pseudo-outcomes (length n).
}
\description{
Compute Doubly Robust Pseudo-Outcomes from Kennedy et al. (2017) JRSS-B
}
