% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_gps.R
\name{gps_model}
\alias{gps_model}
\title{Fit a Global Propensity Score Model f(X | C)}
\usage{
gps_model(X, C, pi_fitter, ...)
}
\arguments{
\item{X}{Numeric matrix or data frame of observed treatments.}

\item{C}{Numeric matrix or data frame of observed confounders.}

\item{pi_fitter}{Function(X_df, C_df, ...) that trains and returns a density model.}

\item{...}{Additional arguments passed to the pi_fitter.}
}
\value{
An S3 object of class "gps_model".
}
\description{
Fit a Global Propensity Score Model f(X | C)
}
