% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_gps.R
\name{mvn_gps_fitter}
\alias{mvn_gps_fitter}
\title{Inner Fitter for MVN GPS}
\usage{
mvn_gps_fitter(
  A = NULL,
  C,
  X = NULL,
  method_gps = c("linear", "SuperLearner"),
  delta_n = 1e-16,
  ...
)
}
\arguments{
\item{A}{Numeric matrix or data frame of observed treatments.}

\item{C}{Numeric matrix or data frame of observed confounders.}

\item{X}{Compatibility alias for \code{A}.}

\item{method_gps}{String in \code{c("linear","SuperLearner")}.}

\item{...}{Additional arguments passed to the density fitter.}
}
\description{
Inner Fitter for MVN GPS
}
