% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_gps.R
\name{predict.gps_model}
\alias{predict.gps_model}
\title{Predict Method for GPS Model}
\usage{
\method{predict}{gps_model}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class "gps_model".}

\item{newdata}{A data frame containing predictors for the nuisance models.}

\item{...}{Additional arguments passed to the inner predict method.}
}
\description{
Predict Method for GPS Model
}
