% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_gps.R
\name{predict.mvn_inner}
\alias{predict.mvn_inner}
\title{Predict Method for Inner MVN}
\usage{
\method{predict}{mvn_inner}(object, newdata, delta_n = 1e-16, ...)
}
\description{
Predict Method for Inner MVN
}
