% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_outcome_C.R
\name{predict.nuisance_C_model}
\alias{predict.nuisance_C_model}
\title{Predict Method for Nuisance C Model}
\usage{
\method{predict}{nuisance_C_model}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class "nuisance_C_model".}

\item{newdata}{A data frame containing the confounders (C).}

\item{...}{Additional arguments passed to the underlying predict method.}
}
\value{
A numeric vector of predictions.
}
\description{
Predict Method for Nuisance C Model
}
