% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_outcome_regression.R
\name{predict.outcome_model}
\alias{predict.outcome_model}
\title{Predict Method for Outcome Model}
\usage{
\method{predict}{outcome_model}(object, newdata, ...)
}
\description{
Predict Method for Outcome Model
}
