% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\alias{normal_expression}
\alias{sick_expression}
\title{Sample matrices for CSD}
\format{
Numeric matrices of normalized gene expression.
Genes are in columns, whereas samples are in rows.

An object of class \code{matrix} (inherits from \code{array}) with 504 rows and 1000 columns.
}
\source{
For the expression matrix for healthy individuals,
GenotypeTissue Expression (GTEx) V7.
For the thyroid cancer patients,
the data are obtained for the Thyroid Cancer
project (THCA) from The Cancer Genome Atlas
}
\usage{
normal_expression

sick_expression
}
\description{
Sample expression matrices of thyroid gland tissue for
thyroid cancer patients and healthy individuals.
Due to size requirements, only 1000 randomly selected genes
are provided in the dataset. Number of samples are 399 and 504
in the healthy controls and the sick samples, respectively.
}
\keyword{datasets}
