% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csdr.R
\name{scores}
\alias{scores}
\alias{scores.csdr_fit}
\title{Extract CSDR scores}
\usage{
scores(object, ...)

\method{scores}{csdr_fit}(object, variant = NULL, ...)
}
\arguments{
\item{object}{A fitted object.}

\item{...}{Additional arguments passed to methods.}

\item{variant}{Optional variant name. If \code{NULL}, return all variants.}
}
\description{
Extract CSDR scores
}
