% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csdr.R
\name{targets}
\alias{targets}
\alias{targets.csdr_fit}
\title{Extract generated CSDR targets}
\usage{
targets(object, ...)

\method{targets}{csdr_fit}(object, variant = NULL, ...)
}
\arguments{
\item{object}{A fitted object.}

\item{...}{Additional arguments passed to methods.}

\item{variant}{Optional variant name. If \code{NULL}, return all variants.}
}
\description{
Extract generated CSDR targets
}
