% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossfit_ERS.R
\name{crossfit_ERS}
\alias{crossfit_ERS}
\title{Estimate Cross-Fitted Causal Exposure Response Surface (Colangelo and Lee, 2026)}
\usage{
crossfit_ERS(
  Y,
  X,
  C,
  x_eval = NULL,
  estimator = c("DR", "RA", "IPW"),
  L = 5,
  outcome_fitter = SL_outcome_fitter,
  gps_fitter = mvn_fitter,
  seed = 42,
  h = NULL,
  c_multiplier = 1.25,
  delta_n = 1e-16,
  optimize_bw = FALSE,
  args_outcome = list(),
  args_gps = list()
)
}
\arguments{
\item{Y}{Numeric vector of observed outcomes (length n).}

\item{X}{Numeric matrix or data frame of observed treatments (n x p).}

\item{C}{Numeric matrix or data frame of observed confounders (n x q).}

\item{x_eval}{Matrix or data frame of target treatment values to evaluate (m x p).}

\item{estimator}{String indicating the estimator to use: "DR", "RA", or "IPW". Defaults to "DR".}

\item{L}{Integer indicating the number of folds for cross-fitting. Defaults to 5.}

\item{outcome_fitter}{Function to train outcome models. Required for RA and DR.}

\item{gps_fitter}{Function to train GPS models. Required for IPW and DR.}

\item{seed}{Random seed for cross-fitting folds.}

\item{h}{Optional numeric vector of bandwidths (length p). Defaults to rule-of-thumb.}

\item{c_multiplier}{Numeric scalar for rule-of-thumb bandwidth calculation. Defaults to 1.25.}

\item{delta_n}{Small positive threshold to floor propensity scores. Defaults to 1e-16.}

\item{optimize_bw}{Logical. If TRUE and estimator is "DR", dynamically calculates the AMSE-optimal bandwidth.}

\item{args_outcome}{List of additional arguments to pass to outcome_fitter.}

\item{args_gps}{List of additional arguments to pass to gps_fitter.}
}
\value{
A list containing the globally aggregated 'results' data frame and 'metadata'.
}
\description{
Estimate Cross-Fitted Causal Exposure Response Surface (Colangelo and Lee, 2026)
}
