% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csdr.R
\name{csdr}
\alias{csdr}
\title{Causal Sufficient Dimension Reduction}
\usage{
csdr(
  Y,
  X,
  C,
  method = c("DR"),
  args_compute_new_response = list(),
  args_outcome = list(),
  args_C = list(),
  args_gps = list(),
  args_ers = list(),
  args_MAVE = list()
)
}
\arguments{
\item{Y}{Numeric vector of outcomes (length n).}

\item{X}{Numeric matrix or data frame of observed treatments (n x p).}

\item{C}{Numeric matrix or data frame of observed confounders (n x q).}

\item{method}{String or vector indicating which causal transformation(s) to use ("RA", "DR", "PO", "RP").}

\item{args_compute_new_response}{List of top-level arguments for \code{compute_new_response_and_exposure}
(e.g., list(L = 10, outcome_fitter = SL_outcome_fitter, seed = 123)).}

\item{args_outcome}{List of tuning parameters for the outcome fitter.}

\item{args_C}{List of tuning parameters for the C fitter.}

\item{args_gps}{List of tuning parameters for the GPS fitter.}

\item{args_ers}{List of tuning parameters for the ERS estimation.}

\item{args_MAVE}{List of arguments to pass to \code{MAVE::mave} (e.g., list(method = "meanOPG")).}
}
\value{
A list containing the MAVE fit, dimension estimation object, selected dimension,
the generated pseudo-data, and run metadata. If multiple methods are requested,
returns a named list of such objects.
}
\description{
Causal Sufficient Dimension Reduction
}
