% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners.R
\name{csdr_learners}
\alias{csdr_learners}
\title{Specify learners for CSDR nuisance fitting}
\usage{
csdr_learners(
  outcome = NULL,
  gps = NULL,
  rp_y = NULL,
  rp_a = NULL,
  sl_library = NULL
)
}
\arguments{
\item{outcome}{Learner for \code{E[Y | A, C]}.}

\item{gps}{Learner for \code{f(A | C)}.}

\item{rp_y}{Learner for \code{E[Y | C]} in residualized-pair targets.}

\item{rp_a}{Learner for \code{E[A_j | C]} in residualized-pair targets.}

\item{sl_library}{Optional SuperLearner library used by unspecified
SuperLearner regression learners.}
}
\value{
A list of CSDR learner specifications with class \code{"csdr_learners"}.
}
\description{
Specify learners for CSDR nuisance fitting
}
