% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_residualized_pair.R
\name{estimate_residualized_pair}
\alias{estimate_residualized_pair}
\title{Compute Residualized Pair}
\usage{
estimate_residualized_pair(Y, X, C, C_models)
}
\arguments{
\item{Y}{Numeric vector of observed outcomes (length n).}

\item{X}{Numeric matrix or data frame of observed treatments (n x p).}

\item{C}{Numeric matrix or data frame of observed confounders (n x q).}

\item{C_models}{A list containing pre-fitted nuisance models:
\code{C_models$Y_model} (predicts E(Y | C)) and
\code{C_models$X_models} (a list of p models predicting E(X_i | C)).}
}
\value{
A list containing two objects: \code{Ytilde} (numeric vector) and \code{Xtilde} (numeric matrix).
}
\description{
Compute Residualized Pair
}
