% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuisance_gps.R
\name{predict.mvn_inner}
\alias{predict.mvn_inner}
\title{Predict Method for Inner MVN}
\usage{
\method{predict}{mvn_inner}(object, newdata, delta_n = 1e-16, ...)
}
\arguments{
\item{object}{An object of class "mvn_inner".}

\item{newdata}{A data frame containing predictors.}

\item{delta_n}{Floor for density predictions.}

\item{...}{Additional arguments passed to the inner predict method.}
}
\description{
Predict Method for Inner MVN
}
