% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_residualized_pair.R
\name{train_nuisance_models}
\alias{train_nuisance_models}
\title{Train all Nuisance Models E(Y | C) and E(X_i | C)}
\usage{
train_nuisance_models(Y, X, C, fitter, ...)
}
\arguments{
\item{Y}{Numeric vector of outcomes.}

\item{X}{Numeric matrix of treatments.}

\item{C}{Numeric matrix of confounders.}

\item{fitter}{The wrapper function to train the models (e.g., SL_nuisance_fitter).}

\item{...}{Additional arguments passed to the fitter.}
}
\value{
A list containing $Y_model and $X_models.
}
\description{
Train all Nuisance Models E(Y | C) and E(X_i | C)
}
